client与server分开
