+++ nacos-encryption-plugin-ext/nacos-aes-encryption-server/pom.xml
@@ -3,29 +3,19 @@
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <parent>
-        <artifactId>nacos-plugin-ext</artifactId>
+        <artifactId>nacos-encryption-plugin-ext</artifactId>
         <groupId>com.alibaba.nacos</groupId>
         <version>${revision}</version>
     </parent>
     <modelVersion>4.0.0</modelVersion>
 
-    <artifactId>nacos-encryption-plugin-ext</artifactId>
-    <packaging>pom</packaging>
-
-    <modules>
-        <module>nacos-aes-encryption-plugin</module>
-        <module>nacos-aes-encryption-client</module>
-        <module>nacos-aes-encryption-server</module>
-    </modules>
-
+    <artifactId>nacos-aes-encryption-plugin-server</artifactId>
+    
     <dependencies>
         <dependency>
             <groupId>com.alibaba.nacos</groupId>
-            <artifactId>nacos-encryption-plugin</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>com.alibaba.nacos</groupId>
-            <artifactId>nacos-common</artifactId>
+            <artifactId>nacos-aes-encryption-plugin</artifactId>
+            <version>${revision}</version>
         </dependency>
     </dependencies>
